backend/requirement.txt
amqp==5.3.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
Faker==37.4.2
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.4
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
